+++ src/main/java/com/onenth/OneNth/domain/member/settings/alert/generalAlert/service/GeneralAlertCommandService.java
@@ -20,8 +20,4 @@ public interface GeneralAlertCommandService {
             GeneralAlertRequestDTO.SetChatAlertStatusRequestDTO request
     );
 
-    GeneralAlertResponseDTO.GetAllAlertSettingsResponseDTO getAllAlertSettings(
-            Long userId
-    );
-
 }
